Factor districts selection into reusable component.
